backend/requirements.txt
django>=5.1,<7.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2
